Moved to configRouting.js
